# models/auth.go
package models

type Authentication struct {
	Email    string `json:"email" validate:"required,email"`
	Password string `json:"password" validate:"required"`
}

type Token struct {
	Role        string `json:"role"`
	Email       string `json:"email"`
	TokenString string `json:"token"`
}

type JsonResponse struct {
	Message string `json:"message"`
	Data    Post   `json:"data"`
}